requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Geospatial Libraries
geopandas==0.13.2
shapely==2.0.1
folium==0.14.0
h3==3.7.6

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
geoalchemy2==0.14.0

# MLOps and Experiment Tracking
mlflow==2.5.0
optuna==3.3.0

# API Framework
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic
python-multipart==0.0.6

# Data Validation
great-expectations==0.17.12

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.65.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Date/Time Handling
python-dateutil==2.8.2
holidays==0.29

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Performance
numba==0.57.1

# Monitoring
prometheus-client==0.17.1

# Logging
loguru==0.7.0
